// app/interfaces/Product.d.ts
export interface Product {
  id: string;
  name: string;
  description: string;
  logo: string;
  date_release: Date;
  date_revision: Date;
}

export interface UpdateProductInput {
  id: string;
  name?: string;
  description?: string;
  logo?: string;
  date_release?: Date;
  date_revision?: Date;
}